Finish playlist implementation
Also:
- Allow for custom reactions on paginations with custom callbacks
- Bug fixes with pagination

+++ EduardoBotv2.Core/Modules/Games/Database/DatabasePokemonRepository.cs
@@ -16,11 +16,11 @@ namespace EduardoBotv2.Core.Modules.Games.Database
             _connectionString = credentials.DbConnectionString;
         }
 
-        public async Task<Dictionary<PokemonSummary, int>> GetPokemonAsync(ulong discordUserId, ulong guildId)
+        public async Task<Dictionary<PokemonSummary, int>> GetPokemonAsync(long discordUserId, long guildId)
         {
             AsyncDataReader dr = new AsyncDataReader("POKEMON_GetPokemonInventory", _connectionString);
-            dr.AddParameter("@DiscordGuildId", (long)guildId);
-            dr.AddParameter("@DiscordUserId", (long)discordUserId);
+            dr.AddParameter("@DiscordGuildId", guildId);
+            dr.AddParameter("@DiscordUserId", discordUserId);
 
             Dictionary<PokemonSummary, int> pokemon = new Dictionary<PokemonSummary, int>();
             await dr.ExecuteReaderAsync(reader =>
@@ -34,11 +34,11 @@ namespace EduardoBotv2.Core.Modules.Games.Database
             return pokemon;
         }
 
-        public async Task AddPokemonAsync(ulong discordUserId, ulong guildId, PokemonSummary pokemon)
+        public async Task AddPokemonAsync(long discordUserId, long guildId, PokemonSummary pokemon)
         { 
             AsyncDataReader dr = new AsyncDataReader("POKEMON_AddPokemonToInventory", _connectionString);
-            dr.AddParameter("@DiscordGuildId", (long)guildId);
-            dr.AddParameter("@DiscordUserId", (long)discordUserId);
+            dr.AddParameter("@DiscordGuildId", guildId);
+            dr.AddParameter("@DiscordUserId", discordUserId);
             dr.AddParameter("@PokemonNumber", pokemon.Id);
             dr.AddParameter("@PokemonName", pokemon.Name);
